feat/del: moved deleteMemberFromProject from member_project to member module
